Add Support for Firefox (and switch to manifest v2)
Added support for Firefox by downgrading to manifest v2. However,
this has no real effect other than changing how the background
page is declared (background script vs service worker).

To get the script working on Firefox, it was necessary to tweak
a few calls in the foreground script, but they are now marked,
and the extension works in both Firefox and Chrome.